refactor: fix incorrect import paths
