+++ imports/plugins/core/cart/server/no-meteor/mutations/setEmailOnAnonymousCart.js
@@ -1,6 +1,6 @@
 import SimpleSchema from "simpl-schema";
 import ReactionError from "@reactioncommerce/reaction-error";
-import hashLoginToken from "/imports/plugins/core/accounts/server/no-meteor/util/hashLoginToken";
+import hashLoginToken from "/imports/node-app/core/util/hashLoginToken";
 
 const inputSchema = new SimpleSchema({
   cartId: String,
